+++ examples/example-app-router-playground/.storybook/main.ts
@@ -1,0 +1,12 @@
+import type {StorybookConfig} from '@storybook/nextjs';
+
+const config: StorybookConfig = {
+  stories: ['../src/**/*.stories.tsx'],
+  addons: ['storybook-next-intl'],
+  framework: {
+    name: '@storybook/nextjs',
+    options: {}
+  },
+  staticDirs: ['../public']
+};
+export default config;