+++ packages/next-intl/src/navigation/shared/utils.test.tsx
@@ -24,14 +24,98 @@ describe('serializeSearchParams', () => {
 });
 
 describe('compileLocalizedPathname', () => {
-  it('throws when params were not resolved', () => {
-    const locales: ReadonlyArray<string> = ['en'];
-    expect(() =>
-      // @ts-expect-error -- Purposefully miss a param
+  const locales: ReadonlyArray<string> = ['en', 'de'];
+  const pathnames = {
+    '/about/[param]': {
+      en: '/about/[param]',
+      de: '/ueber-uns/[param]'
+    },
+    '/test/[one]/[two]': '/test/[one]/[two]',
+    '/test/[one]/[one]': '/test/[one]/[one]',
+    '/test/[...params]': '/test/[...params]',
+    '/test/[[...params]]': '/test/[[...params]]'
+  } as const;
+
+  it('compiles a pathname that differs by locale', () => {
+    expect(
+      compileLocalizedPathname<typeof locales, '/about/[param]'>({
+        locale: 'en',
+        pathname: '/about/[param]',
+        params: {param: 'value'},
+        pathnames
+      })
+    ).toBe('/about/value');
+    expect(
+      compileLocalizedPathname<typeof locales, '/about/[param]'>({
+        locale: 'de',
+        pathname: '/about/[param]',
+        params: {param: 'wert'},
+        pathnames
+      })
+    ).toBe('/ueber-uns/wert');
+  });
+
+  it('compiles a pathname that is equal for all locales', () => {
+    expect(
       compileLocalizedPathname<typeof locales, '/test/[one]/[two]'>({
         locale: 'en',
         pathname: '/test/[one]/[two]',
-        pathnames: '/test/[one]/[two]',
+        params: {one: '1', two: '2'},
+        pathnames
+      })
+    ).toBe('/test/1/2');
+  });
+
+  it('compiles a pathname where a param appears twice', () => {
+    expect(
+      compileLocalizedPathname<typeof locales, '/test/[one]/[one]'>({
+        locale: 'en',
+        pathname: '/test/[one]/[one]',
+        params: {one: '1'},
+        pathnames
+      })
+    ).toBe('/test/1/1');
+  });
+
+  it('compiles a pathname with a catch-all segment', () => {
+    expect(
+      compileLocalizedPathname<typeof locales, '/test/[...params]'>({
+        locale: 'en',
+        pathname: '/test/[...params]',
+        params: {params: ['a', 'b']},
+        pathnames
+      })
+    ).toBe('/test/a/b');
+  });
+
+  it('compiles a pathname with an optional catch-all segment if the segment is provided', () => {
+    expect(
+      compileLocalizedPathname<typeof locales, '/test/[[...params]]'>({
+        locale: 'en',
+        pathname: '/test/[[...params]]',
+        params: {params: ['a', 'b']},
+        pathnames
+      })
+    ).toBe('/test/a/b');
+  });
+
+  it('compiles a pathname with an optional catch-all segment if the segment is absent', () => {
+    expect(
+      compileLocalizedPathname<typeof locales, '/test/[[...params]]'>({
+        locale: 'en',
+        pathname: '/test/[[...params]]',
+        pathnames
+      })
+    ).toBe('/test');
+  });
+
+  it('throws when params were not resolved', () => {
+    expect(() =>
+      compileLocalizedPathname<typeof locales, '/test/[one]/[two]'>({
+        locale: 'en',
+        pathname: '/test/[one]/[two]',
+        pathnames,
+        // @ts-expect-error -- Purposefully miss a param
         params: {one: '1'}
       })
     ).toThrow(